+++ src/main.rs
@@ -4,41 +4,17 @@ use bevy_inspector_egui::WorldInspectorPlugin;
 pub const HEIGHT: f32 = 720.0;
 pub const WIDTH: f32 = 1280.0;
 
-#[derive(Reflect, Component, Default)]
-#[reflect(Component)]
-pub struct Tower {
-    shooting_timer: Timer,
-    bullet_offset: Vec3,
-}
-
-#[derive(Reflect, Component, Default)]
-#[reflect(Component)]
-pub struct Bullet {
-    direction: Vec3,
-    speed: f32,
-}
-
-#[derive(Reflect, Component, Default)]
-#[reflect(Component)]
-pub struct Lifetime {
-    timer: Timer,
-}
-
-#[derive(Reflect, Component, Default)]
-#[reflect(Component)]
-pub struct Target {
-    speed: f32,
-}
-
-#[derive(Reflect, Component, Default)]
-#[reflect(Component)]
-pub struct Health {
-    value: i32,
-}
-
 pub struct GameAssets {
     bullet_scene: Handle<Scene>,
 }
+
+mod bullet;
+mod target;
+mod tower;
+
+pub use bullet::*;
+pub use target::*;
+pub use tower::*;
 
 fn main() {
     App::new()
@@ -54,21 +30,13 @@ fn main() {
         .add_plugins(DefaultPlugins)
         // Inspector Setup
         .add_plugin(WorldInspectorPlugin::new())
-        .register_type::<Tower>()
-        .register_type::<Lifetime>()
-        .register_type::<Target>()
-        .register_type::<Bullet>()
-        .register_type::<Health>()
         // Our Systems
+        .add_plugin(TowerPlugin)
+        .add_plugin(TargetPlugin)
+        .add_plugin(BulletPlugin)
         .add_startup_system(spawn_basic_scene)
         .add_startup_system(spawn_camera)
         .add_startup_system(asset_loading)
-        .add_system(tower_shooting)
-        .add_system(target_death)
-        .add_system(bullet_collision)
-        .add_system(move_targets)
-        .add_system(move_bullets)
-        .add_system(bullet_despawn)
         .run();
 }
 
@@ -76,96 +44,6 @@ fn asset_loading(mut commands: Commands, assets: Res<AssetServer>) {
     commands.insert_resource(GameAssets {
         bullet_scene: assets.load("Bullet.glb#Scene0"),
     });
-}
-
-fn bullet_despawn(
-    mut commands: Commands,
-    mut bullets: Query<(Entity, &mut Lifetime)>,
-    time: Res<Time>,
-) {
-    for (entity, mut lifetime) in &mut bullets {
-        lifetime.timer.tick(time.delta());
-        if lifetime.timer.just_finished() {
-            commands.entity(entity).despawn_recursive();
-        }
-    }
-}
-
-fn target_death(mut commands: Commands, targets: Query<(Entity, &Health)>) {
-    for (ent, health) in &targets {
-        if health.value <= 0 {
-            commands.entity(ent).despawn_recursive();
-        }
-    }
-}
-
-fn tower_shooting(
-    mut commands: Commands,
-    mut towers: Query<(Entity, &mut Tower, &GlobalTransform)>,
-    targets: Query<&GlobalTransform, With<Target>>,
-    bullet_assets: Res<GameAssets>,
-    time: Res<Time>,
-) {
-    for (tower_ent, mut tower, transform) in &mut towers {
-        tower.shooting_timer.tick(time.delta());
-        if tower.shooting_timer.just_finished() {
-            let bullet_spawn = transform.translation() + tower.bullet_offset;
-
-            let direction = targets
-                .iter()
-                .min_by_key(|target_transform| {
-                    FloatOrd(Vec3::distance(target_transform.translation(), bullet_spawn))
-                })
-                .map(|closest_target| closest_target.translation() - bullet_spawn);
-
-            if let Some(direction) = direction {
-                commands.entity(tower_ent).with_children(|commands| {
-                    commands
-                        .spawn_bundle(SceneBundle {
-                            scene: bullet_assets.bullet_scene.clone(),
-                            transform: Transform::from_translation(tower.bullet_offset),
-                            ..Default::default()
-                        })
-                        .insert(Lifetime {
-                            timer: Timer::from_seconds(1000.5, false),
-                        })
-                        .insert(Bullet {
-                            direction,
-                            speed: 2.5,
-                        })
-                        .insert(Name::new("Bullet"));
-                });
-            }
-        }
-    }
-}
-
-fn move_bullets(mut bullets: Query<(&Bullet, &mut Transform)>, time: Res<Time>) {
-    for (bullet, mut transform) in &mut bullets {
-        transform.translation += bullet.direction.normalize() * bullet.speed * time.delta_seconds();
-    }
-}
-
-fn bullet_collision(
-    mut commands: Commands,
-    bullets: Query<(Entity, &GlobalTransform), With<Bullet>>,
-    mut targets: Query<(&mut Health, &Transform), With<Target>>,
-) {
-    for (bullet, bullet_transform) in &bullets {
-        for (mut health, target_transform) in &mut targets {
-            if Vec3::distance(bullet_transform.translation(), target_transform.translation) < 0.2 {
-                commands.entity(bullet).despawn_recursive();
-                health.value -= 1;
-                break;
-            }
-        }
-    }
-}
-
-fn move_targets(mut targets: Query<(&Target, &mut Transform)>, time: Res<Time>) {
-    for (target, mut transform) in &mut targets {
-        transform.translation.x += target.speed * time.delta_seconds();
-    }
 }
 
 fn spawn_basic_scene(
